Extract HealthKitConfig into separate files
